Unify ONNX/PT2 compilation paths and extract shared helpers
Restructure so both ONNX and PT2 paths follow the same call flow:
  lib.rs (thin PyO3 wrapper)
    → onnx_translator.rs / pt2_compiled_model.rs (format-specific translate + compile)
      → compiled_graph.rs::parse_graph (shared backend pipeline)

Rust changes:
- Create onnx_translator.rs with compile_onnx() and translate_onnx()
  (moved from compiled_graph.rs and lib.rs)
- compiled_graph.rs now only contains shared code (GraphTranslation,
  WeightData, CompiledGraph, parse_graph)
- Cache label_map in CompiledGraph for O(1) set_weight_* lookups
- Move weight_device_ptrs into WeightData.device_ptrs
- Add search_iters param to process_onnx (parity with PT2)
- Fix .unwrap() → ? error propagation in ONNX file loading
- lib.rs reduced to thin PyO3 registration layer

Python changes:
- Extract _collect_weight_pointers(), _detect_backend(),
  _load_cpu_weights() shared helpers in main.py
- Both ONNX and PT2 paths use the same helpers
- Centralize _register_cache_serialization() in __init__.py
- CompiledModel: add input_names override, keep user_indices for
  torch.compile lifted-param filtering
