+++ crates/luminal_python/rust/src/pt2_compiled_model.rs
@@ -57,16 +57,11 @@ fn compile_pt2(
     search_iters: usize,
     weight_device_ptrs: HashMap<String, (u64, usize)>,
 ) -> anyhow::Result<CompiledGraph> {
-    let (translation, weights) = translate_pt2(pt2_path, weights_path)?;
+    let (translation, mut weights) = translate_pt2(pt2_path, weights_path)?;
+    weights.device_ptrs = weight_device_ptrs;
 
-    CompiledGraph::parse_graph(
-        translation,
-        weights,
-        backend,
-        search_iters,
-        weight_device_ptrs,
-    )
-    .map_err(|e| anyhow::anyhow!(e))
+    CompiledGraph::parse_graph(translation, weights, backend, search_iters)
+        .map_err(|e| anyhow::anyhow!(e))
 }
 
 /// Translate a PT2 exported model into a format-neutral GraphTranslation + WeightData.
@@ -199,6 +194,7 @@ pub fn translate_pt2(
     let weight_data = WeightData {
         weights,
         tensor_sizes,
+        device_ptrs: HashMap::new(),
     };
 
     Ok((translation, weight_data))
