+++ crates/luminal_python/src/luminal/compiled_model.py
@@ -1,6 +1,6 @@
 """CompiledModel wrapper for the Rust CompiledGraph."""
 
-from typing import List
+from typing import List, Optional
 
 import torch
 
@@ -8,16 +8,19 @@ import torch
 class CompiledModel:
     """Wrapper around CompiledGraph that handles PyTorch tensor conversion."""
 
-    def __init__(self, graph_result, weight_refs=None, user_indices=None):
+    def __init__(self, graph_result, weight_refs=None, input_names=None, user_indices=None):
         """Initialize with a compiled CompiledGraph from Rust.
 
         Args:
-            graph_result: The CompiledGraph from luminal_python.process_onnx() or compile_pt2()
+            graph_result: The CompiledGraph from luminal_python.process_onnx() or process_pt2()
             weight_refs: List of PyTorch tensors to keep alive (prevents GC of shared weights)
-            user_indices: Indices of actual user inputs in __call__ args (None = all are user inputs)
+            input_names: Override for user input names. If None, uses graph_result.input_names.
+            user_indices: When torch.compile lifts model parameters into extra args,
+                this tells __call__ which arg positions are actual user inputs.
+                None means all args are user inputs (PT2 path).
         """
         self._graph = graph_result
-        self._input_names = graph_result.input_names
+        self._input_names = input_names or graph_result.input_names
         self._output_names = graph_result.output_names
         self._output_shapes = graph_result.output_shapes
         self._has_dynamic_dims = getattr(graph_result, 'has_dynamic_dims', False)
@@ -41,22 +44,22 @@ class CompiledModel:
         """Execute the compiled model with PyTorch tensor inputs.
 
         Args:
-            *inputs: PyTorch tensors matching the model's input signature
+            *inputs: PyTorch tensors. When torch.compile lifts model parameters,
+                this includes both weights and user inputs. user_indices filters
+                to just the user inputs.
 
         Returns:
             Tuple of PyTorch tensors containing the model outputs
         """
-        # Determine which inputs are user inputs vs pre-loaded weights
+        # Extract user inputs (torch.compile may pass lifted weights as extra args)
         if self._user_indices is not None:
             user_inputs = [inputs[i] for i in self._user_indices]
-            user_names = [self._input_names[i] for i in self._user_indices]
         else:
             if len(inputs) != len(self._input_names):
                 raise ValueError(
                     f"Expected {len(self._input_names)} inputs, got {len(inputs)}"
                 )
             user_inputs = inputs
-            user_names = self._input_names
 
         input_device = inputs[0].device if inputs else torch.device("cpu")
 
@@ -66,13 +69,11 @@ class CompiledModel:
             self._graph.auto_set_dims_from_input_shapes(input_shapes)
 
         # Set user input data via pointer (avoids Python list conversion)
-        for name, tensor in zip(user_names, user_inputs):
+        for name, tensor in zip(self._input_names, user_inputs):
             if self._is_cuda and tensor.is_cuda:
-                # Zero-copy: pass CUDA device pointer directly
                 t = tensor.detach().contiguous().float()
                 self._graph.set_input_device_ptr(name, t.data_ptr(), t.numel() * 4)
             else:
-                # Copy from host pointer (still avoids .tolist() overhead)
                 t = tensor.detach().cpu().contiguous().float()
                 self._graph.set_input_from_ptr(name, t.data_ptr(), t.numel())
 
@@ -96,5 +97,4 @@ class CompiledModel:
             )
             outputs.append(tensor)
 
-        # Return as a tuple (TorchDynamo expects tuple return from backend callables)
         return tuple(outputs)
