+++ crates/luminal_python/src/luminal/main.py
@@ -1,14 +1,64 @@
 import os
 import tempfile
 
-import onnx
 import torch
 import torch._dynamo
 
 import luminal
 
-from .cache_utils import _register_cache_serialization
 from .compiled_model import CompiledModel
+
+
+# ---------------------------------------------------------------------------
+# Shared helpers (used by both ONNX and PT2 paths)
+# ---------------------------------------------------------------------------
+
+
+def _detect_backend(example_inputs):
+    """Detect backend from input device. Returns 'cuda' or 'cpu'."""
+    device = example_inputs[0].device if example_inputs else torch.device("cpu")
+    return "cuda" if device.type == "cuda" else "cpu"
+
+
+def _collect_weight_pointers(weights, backend):
+    """Partition weight tensors into CUDA device pointers and CPU host pointers.
+
+    Args:
+        weights: dict of name -> torch.Tensor
+        backend: "cuda", "gpu", "cpu", or "native"
+
+    Returns:
+        (keep_alive, device_ptrs, cpu_ptrs) where:
+        - keep_alive: list[Tensor] to prevent GC of shared weight memory
+        - device_ptrs: {name: (device_ptr, n_bytes)}
+        - cpu_ptrs: {name: (host_ptr, n_elements)}
+    """
+    keep_alive = []
+    device_ptrs = {}
+    cpu_ptrs = {}
+    for name, tensor in weights.items():
+        t = tensor.detach().contiguous()
+        if t.dtype != torch.float32:
+            t = t.float()
+        if backend in ("cuda", "gpu") and t.is_cuda:
+            keep_alive.append(t)
+            device_ptrs[name] = (t.data_ptr(), t.numel() * 4)
+        else:
+            t = t.cpu() if t.is_cuda else t
+            keep_alive.append(t)
+            cpu_ptrs[name] = (t.data_ptr(), t.numel())
+    return keep_alive, device_ptrs, cpu_ptrs
+
+
+def _load_cpu_weights(compiled_graph, cpu_weights):
+    """Load CPU weight data into a compiled graph after Rust compilation."""
+    for name, (ptr, n_elements) in cpu_weights.items():
+        compiled_graph.set_weight_from_ptr(name, ptr, n_elements)
+
+
+# ---------------------------------------------------------------------------
+# torch.compile backend entry point
+# ---------------------------------------------------------------------------
 
 
 def luminal_backend(gm, example_inputs, options=None):
@@ -31,45 +81,38 @@ def luminal_backend(gm, example_inputs, options=None):
     )
     opset = options.get("opset", 20)
 
-    _register_cache_serialization()
-    device = example_inputs[0].device if example_inputs else torch.device("cpu")
-    backend = "cuda" if device.type == "cuda" else "native"
+    backend = _detect_backend(example_inputs)
 
     if export_mode == "pt2":
         return _compile_pt2(gm, example_inputs, backend)
     return _compile_onnx(gm, example_inputs, backend, opset=opset)
 
 
+# ---------------------------------------------------------------------------
+# ONNX compilation path
+# ---------------------------------------------------------------------------
+
+
 def _compile_onnx(gm, example_inputs, backend, opset=20):
     """ONNX compilation path."""
     # Identify weight vs user inputs from FX graph placeholders.
     # torch.compile lifts model parameters into graph inputs — we detect them by name prefix.
-    weight_map = {}  # onnx_name -> (placeholder_idx, tensor)
+    weight_tensors = {}  # onnx_name -> tensor
     user_indices = []
     ph_idx = 0
     for node in gm.graph.nodes:
         if node.op == "placeholder":
             onnx_name = f"input_{ph_idx}"
             if node.name.startswith(("l_self_", "l_model_", "l__self_")):
-                weight_map[onnx_name] = (ph_idx, example_inputs[ph_idx])
+                weight_tensors[onnx_name] = example_inputs[ph_idx]
             else:
                 user_indices.append(ph_idx)
             ph_idx += 1
 
-    # Collect weight device pointers upfront so Rust can use them during search
-    # (avoids allocating duplicate GPU buffers from ONNX file data)
-    weight_refs = []
-    weight_device_ptrs = {}  # onnx_name -> (device_ptr, n_bytes)
-    cpu_weights = {}         # onnx_name -> (host_ptr, n_elements)
-    for name, (_idx, tensor) in weight_map.items():
-        if backend == "cuda" and tensor.is_cuda:
-            t = tensor.detach().contiguous().float()
-            weight_refs.append(t)
-            weight_device_ptrs[name] = (t.data_ptr(), t.numel() * 4)
-        else:
-            t = tensor.detach().cpu().contiguous().float()
-            weight_refs.append(t)
-            cpu_weights[name] = (t.data_ptr(), t.numel())
+    # Collect weight pointers for Rust (avoids duplicate GPU buffer allocation)
+    weight_refs, weight_device_ptrs, cpu_weights = _collect_weight_pointers(
+        weight_tensors, backend
+    )
 
     tmp = tempfile.NamedTemporaryFile(suffix=".onnx", delete=False)
     tmp_path = tmp.name
@@ -84,18 +127,24 @@ def _compile_onnx(gm, example_inputs, backend, opset=20):
             input_names=[f"input_{i}" for i in range(len(example_inputs))],
         )
 
-        result = luminal.process_onnx(tmp_path, backend, weight_device_ptrs)
+        result = luminal.process_onnx(tmp_path, backend, weight_device_ptrs=weight_device_ptrs)
     finally:
         os.unlink(tmp_path)
 
-    # Persist all weight inputs and set CPU weights after compilation
-    for name in weight_map:
-        result.persist_input(name)
-    for name, (ptr, n_elements) in cpu_weights.items():
-        result.set_input_from_ptr(name, ptr, n_elements)
+    # Load CPU weights after compilation
+    _load_cpu_weights(result, cpu_weights)
 
-    compiled = CompiledModel(result, weight_refs=weight_refs, user_indices=user_indices)
-    return compiled
+    # Only expose user input names to CompiledModel (weights are pre-loaded).
+    # user_indices tells __call__ which args from torch.compile are real user inputs.
+    user_input_names = [f"input_{i}" for i in user_indices]
+    return CompiledModel(
+        result, weight_refs=weight_refs, input_names=user_input_names, user_indices=user_indices
+    )
+
+
+# ---------------------------------------------------------------------------
+# PT2 compilation path (delegates to pt2 module)
+# ---------------------------------------------------------------------------
 
 
 def _compile_pt2(gm, example_inputs, backend):
